+++ processor/src/test/resources/expected/com/example/inject/Arez_FactoryConsumer3Model.java
@@ -17,7 +17,7 @@ import org.realityforge.braincheck.Guards;
 
 @Generated("arez.processor.ArezProcessor")
 @Singleton
-final class Arez_FactoryConsumer3Model extends FactoryConsumer3Model implements Disposable, Identifiable<Integer> {
+public final class Arez_FactoryConsumer3Model extends FactoryConsumer3Model implements Disposable, Identifiable<Integer> {
   private static volatile int $$arezi$$_nextId;
 
   private final ComponentKernel $$arezi$$_kernel;