+++ processor/src/test/resources/expected/com/example/inject/Arez_FactoryConsumer5Model.java
@@ -16,7 +16,7 @@ import org.realityforge.braincheck.Guards;
 
 @Generated("arez.processor.ArezProcessor")
 @Singleton
-final class Arez_FactoryConsumer5Model extends FactoryConsumer5Model implements Disposable, Identifiable<Integer> {
+public final class Arez_FactoryConsumer5Model extends FactoryConsumer5Model implements Disposable, Identifiable<Integer> {
   private static volatile int $$arezi$$_nextId;
 
   private final ComponentKernel $$arezi$$_kernel;